Revert "make system_bundle_bin_path helper and resolve failing tests for ruby < 2.6"
This reverts commit e63e844bc7444c6a489fcde0dc7011c6c4807edd.

I want to revert this commit because I plan to revert the monkeypatches
to rubygems bin finders, and this is the only spec that breaks when
doing that.

After all the changes to make sure we never rely on the LOAD_PATH for
requiring our own code, I think the change is no longer needed, and the
spec was way more clear before anyways.